requirement.txt
asgiref==3.9.1
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
mysqlclient==2.2.7
PyJWT==2.8.0
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
